Incomplete Child1
Updated child1 styling to fit better in row2 and to better fit the given specifications. Added callback functionality to app.js for child1 and child3. Made progress on updating child1 depending on the target dropdown in row1: the y-axis label now properly updates and componentDidUpdate() is called, but the data of the graph does not update accordingly.